+++ exporters/trace/zipkin/model_test.go
@@ -16,6 +16,7 @@ package zipkin
 
 import (
 	"fmt"
+	"net"
 	"strconv"
 	"testing"
 	"time"
@@ -57,6 +58,7 @@ func TestModelConversion(t *testing.T) {
 			Attributes: []attribute.KeyValue{
 				attribute.Int64("attr1", 42),
 				attribute.String("attr2", "bar"),
+				attribute.Array("attr3", []int{0, 1, 2}),
 			},
 			MessageEvents: []trace.Event{
 				{
@@ -198,6 +200,9 @@ func TestModelConversion(t *testing.T) {
 			Attributes: []attribute.KeyValue{
 				attribute.Int64("attr1", 42),
 				attribute.String("attr2", "bar"),
+				attribute.String("peer.hostname", "test-peer-hostname"),
+				attribute.String("net.peer.ip", "1.2.3.4"),
+				attribute.Int64("net.peer.port", 9876),
 			},
 			MessageEvents: []trace.Event{
 				{
@@ -343,9 +348,8 @@ func TestModelConversion(t *testing.T) {
 					Attributes: nil,
 				},
 			},
-			StatusCode:    codes.Error,
-			StatusMessage: "404, file not found",
-			Resource:      resource,
+			StatusCode: codes.Unset,
+			Resource:   resource,
 		},
 	}
 
@@ -383,10 +387,11 @@ func TestModelConversion(t *testing.T) {
 				},
 			},
 			Tags: map[string]string{
-				"attr1":                   "42",
-				"attr2":                   "bar",
-				"otel.status_code":        "Error",
-				"otel.status_description": "404, file not found",
+				"attr1":            "42",
+				"attr2":            "bar",
+				"attr3":            "[0,1,2]",
+				"otel.status_code": "Error",
+				"error":            "404, file not found",
 			},
 		},
 		// model for span data with no parent
@@ -422,10 +427,10 @@ func TestModelConversion(t *testing.T) {
 				},
 			},
 			Tags: map[string]string{
-				"attr1":                   "42",
-				"attr2":                   "bar",
-				"otel.status_code":        "Error",
-				"otel.status_description": "404, file not found",
+				"attr1":            "42",
+				"attr2":            "bar",
+				"otel.status_code": "Error",
+				"error":            "404, file not found",
 			},
 		},
 		// model for span data of unspecified kind
@@ -461,10 +466,10 @@ func TestModelConversion(t *testing.T) {
 				},
 			},
 			Tags: map[string]string{
-				"attr1":                   "42",
-				"attr2":                   "bar",
-				"otel.status_code":        "Error",
-				"otel.status_description": "404, file not found",
+				"attr1":            "42",
+				"attr2":            "bar",
+				"otel.status_code": "Error",
+				"error":            "404, file not found",
 			},
 		},
 		// model for span data of internal kind
@@ -500,10 +505,10 @@ func TestModelConversion(t *testing.T) {
 				},
 			},
 			Tags: map[string]string{
-				"attr1":                   "42",
-				"attr2":                   "bar",
-				"otel.status_code":        "Error",
-				"otel.status_description": "404, file not found",
+				"attr1":            "42",
+				"attr2":            "bar",
+				"otel.status_code": "Error",
+				"error":            "404, file not found",
 			},
 		},
 		// model for span data of client kind
@@ -527,7 +532,10 @@ func TestModelConversion(t *testing.T) {
 			LocalEndpoint: &zkmodel.Endpoint{
 				ServiceName: "model-test",
 			},
-			RemoteEndpoint: nil,
+			RemoteEndpoint: &zkmodel.Endpoint{
+				IPv4: net.ParseIP("1.2.3.4"),
+				Port: 9876,
+			},
 			Annotations: []zkmodel.Annotation{
 				{
 					Timestamp: time.Date(2020, time.March, 11, 19, 24, 30, 0, time.UTC),
@@ -539,10 +547,13 @@ func TestModelConversion(t *testing.T) {
 				},
 			},
 			Tags: map[string]string{
-				"attr1":                   "42",
-				"attr2":                   "bar",
-				"otel.status_code":        "Error",
-				"otel.status_description": "404, file not found",
+				"attr1":            "42",
+				"attr2":            "bar",
+				"net.peer.ip":      "1.2.3.4",
+				"net.peer.port":    "9876",
+				"peer.hostname":    "test-peer-hostname",
+				"otel.status_code": "Error",
+				"error":            "404, file not found",
 			},
 		},
 		// model for span data of producer kind
@@ -578,10 +589,10 @@ func TestModelConversion(t *testing.T) {
 				},
 			},
 			Tags: map[string]string{
-				"attr1":                   "42",
-				"attr2":                   "bar",
-				"otel.status_code":        "Error",
-				"otel.status_description": "404, file not found",
+				"attr1":            "42",
+				"attr2":            "bar",
+				"otel.status_code": "Error",
+				"error":            "404, file not found",
 			},
 		},
 		// model for span data of consumer kind
@@ -617,10 +628,10 @@ func TestModelConversion(t *testing.T) {
 				},
 			},
 			Tags: map[string]string{
-				"attr1":                   "42",
-				"attr2":                   "bar",
-				"otel.status_code":        "Error",
-				"otel.status_description": "404, file not found",
+				"attr1":            "42",
+				"attr2":            "bar",
+				"otel.status_code": "Error",
+				"error":            "404, file not found",
 			},
 		},
 		// model for span data with no events
@@ -647,10 +658,10 @@ func TestModelConversion(t *testing.T) {
 			RemoteEndpoint: nil,
 			Annotations:    nil,
 			Tags: map[string]string{
-				"attr1":                   "42",
-				"attr2":                   "bar",
-				"otel.status_code":        "Error",
-				"otel.status_description": "404, file not found",
+				"attr1":            "42",
+				"attr2":            "bar",
+				"otel.status_code": "Error",
+				"error":            "404, file not found",
 			},
 		},
 		// model for span data with an "error" attribute set to "false"
@@ -685,10 +696,7 @@ func TestModelConversion(t *testing.T) {
 					Value:     "ev2",
 				},
 			},
-			Tags: map[string]string{
-				"otel.status_code":        "Error",
-				"otel.status_description": "404, file not found",
-			},
+			Tags: nil, // should be omitted
 		},
 	}
 	gottenOutputBatch := toZipkinSpanModels(inputBatch)
@@ -700,7 +708,7 @@ func zkmodelIDPtr(n uint64) *zkmodel.ID {
 	return &id
 }
 
-func Test_toZipkinTags(t *testing.T) {
+func TestTagsTransformation(t *testing.T) {
 	keyValue := "value"
 	doubleValue := 123.456
 	uintValue := int64(123)
@@ -724,21 +732,16 @@ func Test_toZipkinTags(t *testing.T) {
 				},
 			},
 			want: map[string]string{
-				"double":                  fmt.Sprint(doubleValue),
-				"key":                     keyValue,
-				"ok":                      "true",
-				"uint":                    strconv.FormatInt(uintValue, 10),
-				"otel.status_code":        codes.Unset.String(),
-				"otel.status_description": "",
+				"double": fmt.Sprint(doubleValue),
+				"key":    keyValue,
+				"ok":     "true",
+				"uint":   strconv.FormatInt(uintValue, 10),
 			},
 		},
 		{
 			name: "no attributes",
 			data: &tracesdk.SpanSnapshot{},
-			want: map[string]string{
-				"otel.status_code":        codes.Unset.String(),
-				"otel.status_description": "",
-			},
+			want: nil,
 		},
 		{
 			name: "omit-noerror",
@@ -747,10 +750,7 @@ func Test_toZipkinTags(t *testing.T) {
 					attribute.Bool("error", false),
 				},
 			},
-			want: map[string]string{
-				"otel.status_code":        codes.Unset.String(),
-				"otel.status_description": "",
-			},
+			want: nil,
 		},
 		{
 			name: "statusCode",
@@ -763,10 +763,9 @@ func Test_toZipkinTags(t *testing.T) {
 				StatusMessage: statusMessage,
 			},
 			want: map[string]string{
-				"error":                   "true",
-				"key":                     keyValue,
-				"otel.status_code":        codes.Error.String(),
-				"otel.status_description": statusMessage,
+				"error":            statusMessage,
+				"key":              keyValue,
+				"otel.status_code": codes.Error.String(),
 			},
 		},
 		{
@@ -774,10 +773,7 @@ func Test_toZipkinTags(t *testing.T) {
 			data: &tracesdk.SpanSnapshot{
 				InstrumentationLibrary: instrumentation.Library{},
 			},
-			want: map[string]string{
-				"otel.status_code":        codes.Unset.String(),
-				"otel.status_description": "",
-			},
+			want: nil,
 		},
 		{
 			name: "instrLib-noversion",
@@ -788,9 +784,7 @@ func Test_toZipkinTags(t *testing.T) {
 				},
 			},
 			want: map[string]string{
-				"otel.instrumentation_library.name": instrLibName,
-				"otel.status_code":                  codes.Unset.String(),
-				"otel.status_description":           "",
+				"otel.library.name": instrLibName,
 			},
 		},
 		{
@@ -803,16 +797,154 @@ func Test_toZipkinTags(t *testing.T) {
 				},
 			},
 			want: map[string]string{
-				"otel.instrumentation_library.name":    instrLibName,
-				"otel.instrumentation_library.version": instrLibVersion,
-				"otel.status_code":                     codes.Unset.String(),
-				"otel.status_description":              "",
+				"otel.library.name":    instrLibName,
+				"otel.library.version": instrLibVersion,
 			},
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			got := toZipkinTags(tt.data)
+			if diff := cmp.Diff(got, tt.want); diff != "" {
+				t.Errorf("Diff%v", diff)
+			}
+		})
+	}
+}
+
+func TestRemoteEndpointTransformation(t *testing.T) {
+	tests := []struct {
+		name string
+		data *tracesdk.SpanSnapshot
+		want *zkmodel.Endpoint
+	}{
+		{
+			name: "nil-not-applicable",
+			data: &tracesdk.SpanSnapshot{
+				SpanKind:   trace.SpanKindClient,
+				Attributes: []attribute.KeyValue{},
+			},
+			want: nil,
+		},
+		{
+			name: "nil-not-found",
+			data: &tracesdk.SpanSnapshot{
+				SpanKind: trace.SpanKindConsumer,
+				Attributes: []attribute.KeyValue{
+					attribute.String("attr", "test"),
+				},
+			},
+			want: nil,
+		},
+		{
+			name: "peer-service-rank",
+			data: &tracesdk.SpanSnapshot{
+				SpanKind: trace.SpanKindProducer,
+				Attributes: []attribute.KeyValue{
+					semconv.PeerServiceKey.String("peer-service-test"),
+					semconv.NetPeerNameKey.String("peer-name-test"),
+					semconv.HTTPHostKey.String("http-host-test"),
+				},
+			},
+			want: &zkmodel.Endpoint{
+				ServiceName: "peer-service-test",
+			},
+		},
+		{
+			name: "http-host-rank",
+			data: &tracesdk.SpanSnapshot{
+				SpanKind: trace.SpanKindProducer,
+				Attributes: []attribute.KeyValue{
+					semconv.HTTPHostKey.String("http-host-test"),
+					semconv.DBNameKey.String("db-name-test"),
+				},
+			},
+			want: &zkmodel.Endpoint{
+				ServiceName: "http-host-test",
+			},
+		},
+		{
+			name: "db-name-rank",
+			data: &tracesdk.SpanSnapshot{
+				SpanKind: trace.SpanKindProducer,
+				Attributes: []attribute.KeyValue{
+					attribute.String("foo", "bar"),
+					semconv.DBNameKey.String("db-name-test"),
+				},
+			},
+			want: &zkmodel.Endpoint{
+				ServiceName: "db-name-test",
+			},
+		},
+		{
+			name: "peer-hostname-rank",
+			data: &tracesdk.SpanSnapshot{
+				SpanKind: trace.SpanKindProducer,
+				Attributes: []attribute.KeyValue{
+					keyPeerHostname.String("peer-hostname-test"),
+					keyPeerAddress.String("peer-address-test"),
+					semconv.HTTPHostKey.String("http-host-test"),
+					semconv.DBNameKey.String("http-host-test"),
+				},
+			},
+			want: &zkmodel.Endpoint{
+				ServiceName: "peer-hostname-test",
+			},
+		},
+		{
+			name: "peer-address-rank",
+			data: &tracesdk.SpanSnapshot{
+				SpanKind: trace.SpanKindProducer,
+				Attributes: []attribute.KeyValue{
+					keyPeerAddress.String("peer-address-test"),
+					semconv.HTTPHostKey.String("http-host-test"),
+					semconv.DBNameKey.String("http-host-test"),
+				},
+			},
+			want: &zkmodel.Endpoint{
+				ServiceName: "peer-address-test",
+			},
+		},
+		{
+			name: "net-peer-invalid-ip",
+			data: &tracesdk.SpanSnapshot{
+				SpanKind: trace.SpanKindProducer,
+				Attributes: []attribute.KeyValue{
+					semconv.NetPeerIPKey.String("INVALID"),
+				},
+			},
+			want: nil,
+		},
+		{
+			name: "net-peer-ipv6-no-port",
+			data: &tracesdk.SpanSnapshot{
+				SpanKind: trace.SpanKindProducer,
+				Attributes: []attribute.KeyValue{
+					semconv.NetPeerIPKey.String("0:0:1:5ee:bad:c0de:0:0"),
+				},
+			},
+			want: &zkmodel.Endpoint{
+				IPv6: net.ParseIP("0:0:1:5ee:bad:c0de:0:0"),
+			},
+		},
+		{
+			name: "net-peer-ipv4-port",
+			data: &tracesdk.SpanSnapshot{
+				SpanKind: trace.SpanKindProducer,
+				Attributes: []attribute.KeyValue{
+					semconv.NetPeerIPKey.String("1.2.3.4"),
+					semconv.NetPeerPortKey.Int(9876),
+				},
+			},
+			want: &zkmodel.Endpoint{
+				IPv4: net.ParseIP("1.2.3.4"),
+				Port: 9876,
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := toZipkinRemoteEndpoint(tt.data)
 			if diff := cmp.Diff(got, tt.want); diff != "" {
 				t.Errorf("Diff%v", diff)
 			}